Se arregla problema al eliminar producto del carrito

+++ server/controllers/carController.js
@@ -38,12 +38,22 @@ const carController={
     deleteProduct:async (req,res)=>{
         try {
 
-            console.log(req); return;
+            console.log(req.params);
 
             //Buscamos el registro del carrito por id
             const finder=await Car.findOne({
+                where:{
                     id:req.params.id
+                }    
             });
+
+            if(req.user.id!==finder.dataValues.userId){
+                console.log(req.user.id);
+                console.log(finder.dataValues.userId);
+                return res.status(401).json({
+                    msg: 'No autorizado para esta acción'
+                });
+            }
 
             //eliminamos el registro
             await Car.destroy({where:{ id: finder.dataValues.id }});
